Reverting integer for field time_to_expire

+++ plugins/inputs/ssl/check_ssl.go
@@ -87,7 +87,7 @@ func (c *CheckExpire) Gather(acc telegraf.Accumulator) error {
 				errMessage = errors.New("Warning: Certificate is not being verified")
 				timeToExpire = certs[0].NotAfter.Sub(timeNow)
 			}
-			fields := map[string]interface{}{"time_to_expire": int64(timeToExpire.Seconds()), "error": errMessage}
+			fields := map[string]interface{}{"time_to_expire": timeToExpire.Seconds(), "error": errMessage}
 			// Add metrics
 			acc.AddFields("ssl_cert", fields, tags)
 		}
